+++ demo/admin/src/products/ManufacturersGrid.tsx
@@ -1,6 +1,7 @@
 import { useApolloClient, useQuery } from "@apollo/client";
 import {
     CrudContextMenu,
+    filterByFragment,
     GridFilterButton,
     MainContent,
     muiGridFilterToGql,
@@ -26,7 +27,6 @@ import {
     GQLManufacturersListQuery,
     GQLManufacturersListQueryVariables,
 } from "@src/products/ManufacturersGrid.generated";
-import { filter } from "graphql-anywhere";
 import gql from "graphql-tag";
 import * as React from "react";
 import { FormattedMessage, useIntl } from "react-intl";
@@ -151,7 +151,7 @@ export function ManufacturersGrid() {
                             }}
                             refetchQueries={["ManufacturersList"]}
                             copyData={() => {
-                                return filter(manufacturersFragment, params.row);
+                                return filterByFragment(manufacturersFragment, params.row);
                             }}
                         />
                     </>