+++ packages/admin/cms-admin/src/generator/future/generator.ts
@@ -7,6 +7,7 @@ import { basename, dirname } from "path";
 
 import { generateForm } from "./generateForm";
 import { generateGrid } from "./generateGrid";
+import { UsableFields } from "./generateGrid/usableFields";
 import { writeGenerated } from "./utils/writeGenerated";
 
 type ImportReference = {
@@ -28,9 +29,9 @@ export type FormFieldConfig<T> = (
 export type FormConfig<T extends { __typename?: string }> = {
     type: "form";
     gqlType: T["__typename"];
+    mode?: "edit" | "add" | "all";
     fragmentName?: string;
     fields: FormFieldConfig<T>[];
-    title?: string;
 };
 
 export type TabsConfig = { type: "tabs"; tabs: { name: string; content: GeneratorConfig }[] };
@@ -45,11 +46,12 @@ export type GridColumnConfig<T> = (
     | { type: "dateTime" }
     | { type: "staticSelect"; values?: string[] }
     | { type: "block"; block: ImportReference }
-) & { name: keyof T } & DataGridSettings;
+) & { name: UsableFields<T> } & DataGridSettings;
 export type GridConfig<T extends { __typename?: string }> = {
     type: "grid";
     gqlType: T["__typename"];
     fragmentName?: string;
+    query?: string;
     columns: GridColumnConfig<T>[];
     add?: boolean;
     edit?: boolean;
@@ -57,6 +59,7 @@ export type GridConfig<T extends { __typename?: string }> = {
     copyPaste?: boolean;
     readOnly?: boolean;
     filterProp?: boolean;
+    toolbar?: boolean;
 };
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -64,13 +67,13 @@ export type GeneratorConfig = FormConfig<any> | GridConfig<any> | TabsConfig;
 
 export type GeneratorReturn = { code: string; gqlDocuments: Record<string, string> };
 
-export async function runFutureGenerate() {
+export async function runFutureGenerate(filePattern = "src/**/*.cometGen.ts") {
     const schema = await loadSchema("./schema.gql", {
         loaders: [new GraphQLFileLoader()],
     });
     const gqlIntrospection = introspectionFromSchema(schema);
 
-    const files = await glob("src/**/*.cometGen.ts");
+    const files: string[] = await glob(filePattern);
     for (const file of files) {
         let outputCode = "";
         let gqlDocumentsOutputCode = "";