+++ storybook/src/admin-date-time/DatePicker.tsx
@@ -24,7 +24,7 @@ const Story = () => {
                         <Card>
                             <CardContent>
                                 <Field name="dateOne" label="Date" fullWidth component={FinalFormDatePicker} />
-                                <Field name="dateTwo" label="Clearable" fullWidth clearable component={FinalFormDatePicker} />
+                                <Field name="dateTwo" label="Required" fullWidth required component={FinalFormDatePicker} />
                             </CardContent>
                         </Card>
                         <pre>{JSON.stringify(values, null, 4)}</pre>