+++ packages/admin/admin-date-time/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @comet/admin-date-time
 
+## 6.9.0
+
+### Minor Changes
+
+-   e85837a17: Loosen peer dependency on `react-intl` to allow using v6
+
+### Patch Changes
+
+-   Updated dependencies [9ff9d66c6]
+-   Updated dependencies [e85837a17]
+    -   @comet/admin@6.9.0
+    -   @comet/admin-icons@6.9.0
+
+## 6.8.0
+
+### Patch Changes
+
+-   @comet/admin@6.8.0
+-   @comet/admin-icons@6.8.0
+
+## 6.7.0
+
+### Patch Changes
+
+-   @comet/admin@6.7.0
+-   @comet/admin-icons@6.7.0
+
+## 6.6.2
+
+### Patch Changes
+
+-   @comet/admin@6.6.2
+-   @comet/admin-icons@6.6.2
+
+## 6.6.1
+
+### Patch Changes
+
+-   @comet/admin@6.6.1
+-   @comet/admin-icons@6.6.1
+
+## 6.6.0
+
+### Patch Changes
+
+-   Updated dependencies [95b97d768]
+-   Updated dependencies [6b04ac9a4]
+    -   @comet/admin@6.6.0
+    -   @comet/admin-icons@6.6.0
+
 ## 6.5.0
 
 ### Patch Changes