+++ packages/admin/cms-admin/CHANGELOG.md
@@ -1,5 +1,110 @@
 # @comet/cms-admin
 
+## 6.9.0
+
+### Minor Changes
+
+-   e85837a17: Loosen peer dependency on `react-intl` to allow using v6
+
+### Patch Changes
+
+-   Updated dependencies [9ff9d66c6]
+-   Updated dependencies [8fb8b209a]
+-   Updated dependencies [e85837a17]
+    -   @comet/admin@6.9.0
+    -   @comet/admin-rte@6.9.0
+    -   @comet/admin-date-time@6.9.0
+    -   @comet/blocks-admin@6.9.0
+    -   @comet/admin-icons@6.9.0
+    -   @comet/admin-theme@6.9.0
+
+## 6.8.0
+
+### Patch Changes
+
+-   c1ca9c335: Don't remove references to `DamFile` from blocks when copying a document from one scope to another if DAM scoping is not enabled
+-   Updated dependencies [90c6f192e]
+-   Updated dependencies [90c6f192e]
+    -   @comet/blocks-admin@6.8.0
+    -   @comet/admin@6.8.0
+    -   @comet/admin-date-time@6.8.0
+    -   @comet/admin-icons@6.8.0
+    -   @comet/admin-rte@6.8.0
+    -   @comet/admin-theme@6.8.0
+
+## 6.7.0
+
+### Patch Changes
+
+-   2db3bc855: Fix `CurrentUserInterface` type
+
+    Add missing `id` field, make `name`, `email`, and `language` required.
+
+    -   @comet/admin@6.7.0
+    -   @comet/admin-date-time@6.7.0
+    -   @comet/admin-icons@6.7.0
+    -   @comet/admin-rte@6.7.0
+    -   @comet/admin-theme@6.7.0
+    -   @comet/blocks-admin@6.7.0
+
+## 6.6.2
+
+### Patch Changes
+
+-   0758d2339: Hide the "Dependents" tab in the DAM if the `DependenciesConfigProvider` is not configured
+
+    Previously, the tab was always shown, even if the feature wasn't configured. Though it didn't cause an error, the tab showed no valuable information.
+
+    Now, we hide the tab if no configuration is passed via the `DependenciesConfigProvider`.
+
+    -   @comet/admin@6.6.2
+    -   @comet/admin-date-time@6.6.2
+    -   @comet/admin-icons@6.6.2
+    -   @comet/admin-rte@6.6.2
+    -   @comet/admin-theme@6.6.2
+    -   @comet/blocks-admin@6.6.2
+
+## 6.6.1
+
+### Patch Changes
+
+-   @comet/admin@6.6.1
+-   @comet/admin-date-time@6.6.1
+-   @comet/admin-icons@6.6.1
+-   @comet/admin-rte@6.6.1
+-   @comet/admin-theme@6.6.1
+-   @comet/blocks-admin@6.6.1
+
+## 6.6.0
+
+### Patch Changes
+
+-   c76666503: Make headers in `includeInvisibleContentContext` overridable in query
+
+    You can now override the headers `x-include-invisible-content` and `x-preview-dam-urls` in your query like this:
+
+    ```tsx
+    const { loading, data, error } = useQuery(exampleQuery, {
+        // ...
+        context: {
+            headers: {
+                "x-include-invisible-content": [],
+                "x-preview-dam-urls": 0,
+            },
+        },
+    });
+    ```
+
+-   Updated dependencies [95b97d768]
+-   Updated dependencies [a65679ba3]
+-   Updated dependencies [6b04ac9a4]
+    -   @comet/admin@6.6.0
+    -   @comet/blocks-admin@6.6.0
+    -   @comet/admin-date-time@6.6.0
+    -   @comet/admin-icons@6.6.0
+    -   @comet/admin-rte@6.6.0
+    -   @comet/admin-theme@6.6.0
+
 ## 6.5.0
 
 ### Minor Changes